+++ input-stream/src/main/java/software/amazon/s3/dataaccelerator/util/Constants.java
@@ -24,4 +24,5 @@ public final class Constants {
   public static final int ONE_MB = 1024 * 1024;
   public static final int PARQUET_MAGIC_STR_LENGTH = 4;
   public static final int PARQUET_FOOTER_LENGTH_SIZE = 4;
+  public static final long DEFAULT_MIN_ADJACENT_COLUMN_LENGTH = 500 * ONE_KB;
 }